pyproject.toml
[project]
name = "line-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "fastapi>=0.115.13",
    "form-app",
    "gspread>=6.2.1",
    "line-bot-sdk>=3.17.1",
    "pandas>=2.3.3",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.12.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "shared",
    "uvicorn[standard]>=0.34.3",
]

[tool.uv.workspace]
members = [
    "form_app",
    "flow_charts",
    "shared/database",
    "senders",
    "webhook",
    "shared",
    "alembic"
]

[tool.uv.sources]
shared = { workspace = true }
form-app = { workspace = true }
